onlineshop/requirements.txt
Django~=3.2.7
django-bootstrap4==3.0.1
djangorestframework~=3.11.0
psycopg2>=2.7
django-phonenumber-field[phonenumbers]
Pillow==8.3.2
django-cors-headers